chore: Behandlingsmeny for veiledere
